+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- `collapsedSectionIds` to `OneBlinkUploader.uploadFormSubmissionDraft()`
+- `sectionState` to `OneBlinkUploader.uploadFormSubmissionDraft()`
 
 ## [3.0.0] - 2025-01-12